pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.186.1",
    "flask>=3.1.2",
    "openai>=1.108.0",
    "python-dotenv>=1.1.1",
]
